flowgrind daily commits are back, major bugs fixes to blocksize and rtt handling
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/projects/flowgrind/branches/flowgrind-next@6366 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/common.h
@@ -128,13 +128,12 @@ struct _report
 	struct timeval begin;
 	struct timeval end;
 
-	long long bytes_read;
-	long long bytes_written;
-	/* TODO: figure out if long is needed here */
-	int request_blocks_read;
-	int request_blocks_written;
-	int response_blocks_read;
-	int response_blocks_written;
+	unsigned long long bytes_read;
+	unsigned long long bytes_written;
+	unsigned long request_blocks_read;
+	unsigned long request_blocks_written;
+	unsigned long response_blocks_read;
+	unsigned long response_blocks_written;
 
 	double rtt_min, rtt_max, rtt_sum;
 	double iat_min, iat_max, iat_sum;
